+++ DSUGrupp1/Models/ViewModels/ChartViewModel.cs
@@ -14,7 +14,7 @@ namespace DSUGrupp1.Models.ViewModels
         public Chart Chart { get; set; }
         public string JsonChart { get; set; }
 
-        public Chart CreateChart(string type, List<string> labels, string DatasetLabel, List<int> data, List<string> bgcolor, int bWidth = 5)
+        public Chart CreateChart(string type, List<string> labels, string DatasetLabel, List<double> data, List<string> bgcolor, int bWidth = 5)
         {
             Chart template = new Chart
             {
